# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Caderneta da Professora
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Student {
  id: string;
  name: string;
  createdAt: string;
}

export interface CreateStudentRequest {
  name: string;
}

export interface AttendanceRecord {
  id: string;
  studentId: string;
  date: string;
  present: boolean;
  createdAt: string;
}

export interface UpsertAttendanceRequest {
  studentId: string;
  date: string;
  present: boolean;
}

export type ActivityType = (typeof ActivityType)[keyof typeof ActivityType];

export const ActivityType = {
  homework: "homework",
  classwork: "classwork",
} as const;

export interface Activity {
  id: string;
  subject: string;
  type: ActivityType;
  link?: string;
  date: string;
  description: string;
  createdAt: string;
}

export type CreateActivityRequestType =
  (typeof CreateActivityRequestType)[keyof typeof CreateActivityRequestType];

export const CreateActivityRequestType = {
  homework: "homework",
  classwork: "classwork",
} as const;

export interface CreateActivityRequest {
  subject: string;
  type: CreateActivityRequestType;
  link?: string;
  date: string;
  description: string;
}

export interface DeliveryRecord {
  id: string;
  activityId: string;
  studentId: string;
  delivered: boolean;
  createdAt: string;
}

export interface UpsertDeliveryRequest {
  activityId: string;
  studentId: string;
  delivered: boolean;
}

export type ListAttendanceParams = {
  date?: string;
};

export type ListDeliveriesParams = {
  activityId?: string;
  studentId?: string;
};