# lib/api-zod/src/generated/types/activity.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Caderneta da Professora
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityType } from "./activityType";

export interface Activity {
  id: string;
  subject: string;
  type: ActivityType;
  link?: string;
  date: string;
  description: string;
  createdAt: string;
}