# lib/api-zod/src/generated/types/attendanceRecord.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Caderneta da Professora
 * OpenAPI spec version: 0.1.0
 */

export interface AttendanceRecord {
  id: string;
  studentId: string;
  date: string;
  present: boolean;
  createdAt: string;
}