# lib/api-zod/src/generated/types/createActivityRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Caderneta da Professora
 * OpenAPI spec version: 0.1.0
 */
import type { CreateActivityRequestType } from "./createActivityRequestType";

export interface CreateActivityRequest {
  subject: string;
  type: CreateActivityRequestType;
  link?: string;
  date: string;
  description: string;
}