# lib/api-zod/src/generated/types/createStudentRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Caderneta da Professora
 * OpenAPI spec version: 0.1.0
 */

export interface CreateStudentRequest {
  name: string;
}